backend/pyproject.toml
[tool.poetry]
name = "quant-trading-backend"
version = "0.1.0"
description = "Quantitative Trading System Backend"
authors = ["Your Name <your.email@example.com>"]
packages = [
    { include = "src" }
]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.68.1"
uvicorn = "^0.15.0"
python-binance = "^1.0.15"
pandas = "^1.3.3"
numpy = "^1.21.2"
sqlalchemy = "^1.4.23"
pydantic = "^1.8.2"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
python-multipart = "^0.0.5"
pyyaml = "^6.0.2"
click = "^8.0.0"
psycopg2-binary = "^2.9.1"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
black = "^21.9b0"
flake8 = "^3.9.2"
mypy = "^0.910"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
